expend dtb size in link script
qemu will check the size reserved for qemu by
substract the start of text section and the start
of ram. this size should larger than dtb size, otherwise
dtb will not pass to the vm.
this patch expend dtb size by enlarge the start address of
text in link script.