scan every line of code and give some comments on some of them
as most of these code are migrated from other project,
I will go over all of the code line by line and give comments
for them to study the detail of creating a os for arm64.
this commit can be the beginning of this process. I think I will
complete it quickly.