printf: enable printf
now, printf can use in armos, it will help to debug.
also a bug in gic fixed here. this is a bug from unikraft.
to get the implemented number of interrupti from GICD_TYPE, the mask should
be 0x1f not 0x7.
the exception frame works, as sgi can be signaled and recived. but spi can
not be recived for now. I can't say if the configuration for spi is
woring for gic part or there is really no spi signal. but I think
keyborad signal will generated and route to vm.